refactor(team): regroup type defs + remaining wrappers (C16)
Continued semantic regrouping: each declaration moved to where its
peers live, then the few remaining thin wrappers swept.

Type moves:

* headlessWorkerPool struct -> headless_codex.go (next to
  headlessCodexTurn / headlessCodexActiveTurn — they form a single
  conceptual unit)
* humanInterview.TitleOrDefault method -> broker.go (next to the
  humanInterview type definition; previously orphaned in launcher.go)
* containsSlug helper -> notifier_targets.go (the only in-package
  caller is the routing decision tree)

Method moves:

* (l *Launcher).sendNotificationToPane body -> pane_dispatch.go
  (sits next to launcherSendNotificationToPane which delegates to it
  and the rest of the dispatcher's send path)

Wrapper sweep:

  l.queuePaneNotification(slug, target, notif)  -> l.paneDispatch().Enqueue(...)

queuePaneNotification + runPaneDispatchQueue Launcher wrappers
deleted. Test call sites in pane_dispatch_queue_test.go updated.

launcher.go: 786 -> 728 lines (-58).

Cumulative since main: 4998 -> 728 = -4270 lines (-85.4%).

Stacked on refactor/launcher-c15-semantic-cleanup (C15).
